+++ packages/perseus/src/widgets/radio/__stories__/multiple-choice-option.stories.tsx
@@ -16,7 +16,7 @@ const Container = (props: {children: React.ReactNode}): React.ReactElement => {
 };
 
 export default {
-    title: "Perseus/Widgets/Radio/Choice",
+    title: "Perseus/Widgets/Radio/Widget Internal Components/Choice",
 };
 
 export const SingleSelect = (): React.ReactElement => {
@@ -36,25 +36,27 @@ export const SingleSelect = (): React.ReactElement => {
                 isMultiSelect={false}
                 updateChecked={() => {}}
             >
-                USS Yangtzee Kiang - NCC-72453 - Hijacked by Bajoran terrorist Tahna Los. Became the first DS9 runabout to be destroyed when it crashed on a moon in the Gamma Quadrant.
+                USS Yangtzee Kiang - NCC-72453 - Hijacked by Bajoran terrorist
+                Tahna Los. Became the first DS9 runabout to be destroyed when it
+                crashed on a moon in the Gamma Quadrant.
             </Choice>
             <Choice
                 checked={false}
                 indicatorContent="C"
                 isMultiSelect={false}
-                updateChecked={() => {
-                }}
+                updateChecked={() => {}}
             >
                 <img
-                     src="https://cdn.kastatic.org/ka-content-images/9cb2cf618c16501d01abf97036deb355d9393949.png"/>
+                    alt="triangle"
+                    src="https://cdn.kastatic.org/ka-content-images/9cb2cf618c16501d01abf97036deb355d9393949.png"
+                />
             </Choice>
             <Choice
                 checked={true}
                 indicatorContent="D"
                 isMultiSelect={false}
                 showCorrectness="wrong"
-                updateChecked={() => {
-                }}
+                updateChecked={() => {}}
             >
                 ISS Enterprise - NCC-1701 (Mirror Universe)
             </Choice>
@@ -62,8 +64,7 @@ export const SingleSelect = (): React.ReactElement => {
                 checked={false}
                 indicatorContent="E"
                 isMultiSelect={false}
-                updateChecked={() => {
-                }}
+                updateChecked={() => {}}
             >
                 USS Yamaguchi - NCC-26510
             </Choice>
@@ -72,8 +73,7 @@ export const SingleSelect = (): React.ReactElement => {
                 indicatorContent="F"
                 isMultiSelect={false}
                 showCorrectness="correct"
-                updateChecked={() => {
-                }}
+                updateChecked={() => {}}
             >
                 USS Enterprise - NCC-1701
             </Choice>